fix/MSSDK-1545: extract components based on step
